requirements.txt
numpy<2.0
tensorflow
matplotlib
opencv-python
pillow
gradio